feat: add viz-smol
