+++ viz-smol/src/listener.rs
@@ -1,5 +1,3 @@
-use std::{future::Future, io::Result};
-
 /// A trait for a listener: `TcpListener` and `UnixListener`.
 pub trait Listener {
     /// The stream's type of this listener.
@@ -8,7 +6,9 @@ pub trait Listener {
     type Addr;
 
     /// Accepts a new incoming connection from this listener.
-    fn accept(&self) -> impl Future<Output = Result<(Self::Io, Self::Addr)>> + Send;
+    fn accept(
+        &self,
+    ) -> impl std::future::Future<Output = std::io::Result<(Self::Io, Self::Addr)>> + Send;
 
     /// Returns the local address that this listener is bound to.
     ///
@@ -16,5 +16,5 @@ pub trait Listener {
     ///
     /// An error will return if got the socket address of the local half of this connection is
     /// failed.
-    fn local_addr(&self) -> Result<Self::Addr>;
+    fn local_addr(&self) -> std::io::Result<Self::Addr>;
 }
