+++ viz-smol/src/server/tcp.rs
