--- viz/src/server/tokio/tls.rs
@@ -1,7 +1,0 @@
-/// `native_tls`
-#[cfg(feature = "native-tls")]
-pub mod native_tls;
-
-/// `rustls`
-#[cfg(feature = "rustls")]
-pub mod rustls;
